Add files via upload
This Jupyter Notebook is a comprehensive guide for beginners exploring Hugging Face, a pivotal platform for Large Language Models (LLMs). Taken from Shaw Talebi's educational playlist on Large Language Models, this notebook aims to provide users with a foundational understanding of Hugging Face, its utilities, and how it can be leveraged to develop and deploy state-of-the-art NLP applications.