Add files via upload
The notebook serves as an introductory primer to LangChain, designed for novices in the domain. It elucidates the foundational principles of LangChain, followed by a systematic guide to architecting a solution framework. This framework leverages the OpenAI API for the dynamic generation of a culinary menu predicated on user-specified cuisine preferences. The notebook demystifies integrating LangChain with OpenAI's API through practical implementation, showcasing the procedural steps to configure, invoke, and process API responses for the automated suggestion of cuisine-specific menu items. Inspired by a tutorial by Dhaval Patel.